updates frameworks and busses
This is the start of the work for adding SDL as a display bus. It is going to function like the RGBBus when finished.
I added a display driver to handle SDL using the built in drivers in LVGL. I do not like how this driver works because it doesn't allow the user to handle setting the buffer sizes and things of that nature. It removes all flexability and it could cause an issue if trying to run LVGL on something like an omega onion. I am trying to remove those limitations.