Some Updates to the RGB bus.
RGB Bus is actually working. What I have found out is that you MUST use 2 frame buffers. There is no way around it. The ESP32 is simply not fast enough with MicroPython. If a single frame buffer is used there will be tearing.

With the 2 frame buffers it runs quite well due to the RGB Bus driver being able to use DMA from the SPIRAM. I have to check and see if the I8080 driver will utilize DMA from the SPIRAM. I do not think it will without having to roll my own criver and not use the one that is built into the SDK. This is because the DMA descriptiors need to be allocated at the same time as the structure that holds the buffers and with how the I80 bus works it is not what allocates the memory for the display. This is a feature I really do want to add because it is what is really going to give a sizeable speed boost.