Big update
Adds bounce buffer support to the RGBBus driver.Hopefully this will fix the tearing issues

Changes SPI bus driver to improve performance when using a single buffer.
The driver that was built into the ESP-IDF always used queuing when sending data. This causes ISR's to happen and when using a single buffer the number of ISR's that take place are HUGE. This impacts performance big time. This change should double the speed of the transfers when using only a single buffer.