updates the SPI Bus and also adds another user module
After I have been looking at the code in micropython I decided that it's SPI implementation was going to cause problems is used in conjunction with the SPIBus driver. This would make it impossible to use an SPI based touch driver IC.

MicroPythons implementation of SPI combined the bus and the device into a single class. It would error if the bus was already initialized and that would lead to no way od being able to attach a touch driver using the same bus. This design is simply not going to work at all so I wrote a new SPI driver for the ESP32. It supports dual, quad and octal modes, all the hardware pins and hosts and the biggest thing is it's not going to pitch a fit if the bus has already been initialized. the module is `spi` and I split the bus and device portions apart so you have `spi.SPI` for the SPI bus and you have `spi.Device` for a device. So you only need to create a single `spi.SPI` instance and then for each device you have attached you create an `spi.Device` instance.

Last big improvement of the SPI is it is no longer blocking. It utilizes DMA memory transfers and callbacks to let you know when it has finished transmitting and receiving. You use the device instance to do the sending and receiving and registering the callbacks.

I am going to add stub files for the new SPI module as soon as I know everything is compiling properly with this.